restructure css files
